Split IO/choice interfaces into their own files
